Create module loader to load the page
