+++ src/index.js
@@ -1,25 +1,3 @@
-import { videoLooper } from './video';
-import { drawMenu } from './menuDOM';
-import { contact } from './contactDOM';
-import { landing } from './landingDOM';
-import { navBarDom } from './NavBarDOM';
-import { tabcontent } from './tabContentDOM';
+import { pageLoader } from './load';
 
-const parentAll = document.getElementById('content');
-
-navBarDom.createNav(parentAll);
-tabcontent.createTabs(parentAll);
-
-const parentLanding = document.getElementById('home');
-
-landing.putBackgr(parentLanding);
-
-landing.putPresentation(parentLanding);
-
-const parentContact = document.getElementById('contact-tab');
-
-contact.putForm(parentContact);
-
-drawMenu.addPanelMenu();
-
-videoLooper.videoEndListener();
+pageLoader.pageLoad();
